Fix ESLint errors and build issues: - Create missing CalendarWidget.css - Separate AuthContext and useAuth hook for fast refresh - Fix unused variables in Login and Register components - Fix useEffect dependency warning in Report component - Update imports to use named exports
