+++ src/hooks/useAuth.js
@@ -1,42 +1,6 @@
 import { useContext } from 'react';
-import { useNavigate } from 'react-router-dom';
-import AuthContext from '../contexts/AuthContext';
+import { AuthContext } from '../contexts/AuthContextDef';
 
-const useAuth = () => {
-  const context = useContext(AuthContext);
-  const navigate = useNavigate();
-
-  if (!context) {
-    throw new Error('useAuth must be used within an AuthProvider');
-  }
-
-  const { currentUser, login, logout } = context;
-
-  const handleLogin = async (email, password) => {
-    try {
-      await login(email, password);
-      navigate('/');
-    } catch (error) {
-      console.error('Login failed:', error);
-      throw error;
-    }
-  };
-
-  const handleLogout = async () => {
-    try {
-      await logout();
-      navigate('/login');
-    } catch (error) {
-      console.error('Logout failed:', error);
-    }
-  };
-
-  return {
-    currentUser,
-    isAuthenticated: !!currentUser,
-    login: handleLogin,
-    logout: handleLogout
-  };
-};
-
-export default useAuth;
+export function useAuth() {
+  return useContext(AuthContext);
+}
